working convolve
