+++ hdl/convolve_audio.sv
@@ -37,26 +37,23 @@ module convolve_audio #(parameter IMPULSE_LENGTH = 24000) (
 
     logic [12:0] first_audio_index, second_audio_index;
 
+    // logic signed [15:0] audio_vals_0;
+    // logic signed [15:0] audio_vals_1;
+    // logic signed [15:0] audio_vals_2;
+    // logic signed [15:0] audio_vals_3;
+    // logic signed [15:0] audio_vals_4;
+    // logic signed [15:0] audio_vals_5;
+    // logic signed [15:0] audio_vals_6;
+    // logic signed [15:0] audio_vals_7;
 
-    logic signed [15:0] audio_vals_0;
-    logic signed [15:0] audio_vals_1;
-    logic signed [15:0] audio_vals_2;
-    logic signed [15:0] audio_vals_3;
-    logic signed [15:0] audio_vals_4;
-    logic signed [15:0] audio_vals_5;
-    logic signed [15:0] audio_vals_6;
-    logic signed [15:0] audio_vals_7;
-
-    assign audio_vals_0 = audio_vals[0];
-    assign audio_vals_1 = audio_vals[1];
-    assign audio_vals_2 = audio_vals[2];
-    assign audio_vals_3 = audio_vals[3];
-    assign audio_vals_4 = audio_vals[4];
-    assign audio_vals_5 = audio_vals[5];
-    assign audio_vals_6 = audio_vals[6];
-    assign audio_vals_7 = audio_vals[7];
-
-
+    // assign audio_vals_0 = audio_vals[0];
+    // assign audio_vals_1 = audio_vals[1];
+    // assign audio_vals_2 = audio_vals[2];
+    // assign audio_vals_3 = audio_vals[3];
+    // assign audio_vals_4 = audio_vals[4];
+    // assign audio_vals_5 = audio_vals[5];
+    // assign audio_vals_6 = audio_vals[6];
+    // assign audio_vals_7 = audio_vals[7];
 
     assign audio_vals[0] = last_value_brom0;
     assign audio_vals[2] = last_value_brom1;
