+++ hdl/ir_buffer.sv
@@ -32,9 +32,9 @@ module ir_buffer #(parameter MEMORY_DEPTH = 16'd6000)
         .RAM_WIDTH(16),
         .RAM_DEPTH(MEMORY_DEPTH)
     ) ir_buffer_0 (
-        .addra(write_enable ? (16'd23999 - ir_sample_index) : first_ir_index),
+        .addra((ir_data_in_valid && (ir_sample_index < 6000)) ? (16'd5999 - ir_sample_index) : first_ir_index),
         .clka(audio_clk),
-        .wea(write_enable && ir_data_in_valid && (ir_sample_index >= 16'd18000)),
+        .wea(write_enable && ir_data_in_valid && (ir_sample_index < 6000)),
         .dina(write_data),
         .ena(1'b1),
         .regcea(1'b1),
@@ -55,9 +55,9 @@ module ir_buffer #(parameter MEMORY_DEPTH = 16'd6000)
         .RAM_WIDTH(16),
         .RAM_DEPTH(MEMORY_DEPTH)
     ) ir_buffer_1 (
-        .addra(write_enable ? (16'd17999 - ir_sample_index) : first_ir_index),
+        .addra((ir_data_in_valid && (ir_sample_index < 16'd12000) && (ir_sample_index >= 16'd6000)) ? (16'd11999 - ir_sample_index) : first_ir_index),
         .clka(audio_clk),
-        .wea(write_enable && ir_data_in_valid && (ir_sample_index < 16'd18000) && (ir_sample_index >= 16'd12000)),
+        .wea(write_enable && ir_data_in_valid && (ir_sample_index < 16'd12000) && (ir_sample_index >= 16'd6000)),
         .dina(write_data),
         .ena(1'b1),
         .regcea(1'b1),
@@ -78,9 +78,9 @@ module ir_buffer #(parameter MEMORY_DEPTH = 16'd6000)
         .RAM_WIDTH(16),
         .RAM_DEPTH(MEMORY_DEPTH)
     ) ir_buffer_2 (
-        .addra(write_enable ? (16'd11999 - ir_sample_index) : first_ir_index),
+        .addra((ir_data_in_valid && (ir_sample_index < 16'd18000) && (ir_sample_index >= 16'd12000)) ? (16'd17999 - ir_sample_index) : first_ir_index),
         .clka(audio_clk),
-        .wea(write_enable && ir_data_in_valid && (ir_sample_index < 16'd12000) && (ir_sample_index >= 16'd6000)),
+        .wea(write_enable && ir_data_in_valid && (ir_sample_index < 16'd18000) && (ir_sample_index >= 16'd12000)),
         .dina(write_data),
         .ena(1'b1),
         .regcea(1'b1),
@@ -101,9 +101,9 @@ module ir_buffer #(parameter MEMORY_DEPTH = 16'd6000)
         .RAM_WIDTH(16),
         .RAM_DEPTH(MEMORY_DEPTH)
     ) ir_buffer_3 (
-        .addra(write_enable ? (16'd5999 - ir_sample_index) : first_ir_index),
+        .addra((ir_data_in_valid && (ir_sample_index >= 16'd18000)) ? (16'd23999 - ir_sample_index) : first_ir_index),
         .clka(audio_clk),
-        .wea(write_enable && ir_data_in_valid && (ir_sample_index < 6000)),
+        .wea(write_enable && ir_data_in_valid && (ir_sample_index >= 16'd18000)),
         .dina(write_data),
         .ena(1'b1),
         .regcea(1'b1),
